Refactor ParameterFieldWidgetSelector
