+++ frontend/src/metabase/parameters/components/widgets/ParameterFieldWidget/normalizeValue.js
@@ -1,0 +1,2 @@
+export const normalizeValue = value =>
+  Array.isArray(value) ? value : value != null ? [value] : [];
